% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdatum.R
\name{normalize_coords}
\alias{normalize_coords}
\title{Internal Helper: Standardize Column Names}
\usage{
normalize_coords(coords, req_z = FALSE)
}
\description{
Internal Helper: Standardize Column Names
}
\keyword{internal}
