% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdatum.R
\name{vdatum_regions}
\alias{vdatum_regions}
\title{Get Available Regions for VDatum API}
\usage{
vdatum_regions()
}
\value{
A data frame with region codes and descriptions.
}
\description{
Returns a data frame of available regions and their descriptions.
}
